refactor(puzzle): consolidate bezierPathToSvg / fmt helpers (closes #278)
Move the duplicated helpers into a single canonical
`src/puzzle/composable/bezier-path.ts` and import from the four
former owners:

- `procedural-generator.ts`
- `composable/compose.ts`
- `composable/curve-clamp.ts`
- `fractal-generator.ts` (also had its own `fmt` — folded in)

Picks the integer-optimized `fmt` (drops trailing zeros for whole
numbers) over the always-2-decimal variant in
`procedural-generator.ts`. Visually identical SVG output; path
strings differ only as `0` vs `0.00` for integer coordinates.

`curve-clamp.ts` previously emitted a redundant leading
`L path[0]` from inside `bezierPathToSvg`; the call site already
emits `L pLeft` to mathematically the same point, so dropping it
makes the canonical helper match the other call sites' contract.
